Moved WriteBatch to dedicated module...
